Inline X86Ops.hpp into OMRInstOpCode.hpp
This is in preparation for deprecating `TR_X86OpCodes` in favour of the
cross platform `TR::InstOpCode::Mnemonic` enum.
